Register RURuntimeStats after execution
